* server/stream.{cpp,h} (read_uint): don't force return           to fit in a short. Fixes a regression for 3dtris.swf.

+++ server/stream.cpp
@@ -50,7 +50,7 @@ namespace gnash {
 // 	}
 
 	
-	unsigned short stream::read_uint(unsigned short bitcount)
+	unsigned stream::read_uint(unsigned short bitcount)
 	{
 		assert(bitcount <= 32);
